fix test_forward
