+++ miners/IMAGE/ffnn_grunt/ffnn_grunt.py
@@ -56,7 +56,7 @@ class Miner():
         # ---- Build FFNN Model ----
         self.model = FFNNNucleus( self.config )
         self.model.to(torch.device("cuda" if torch.cuda.is_available() else "cpu"))
-        bittensor.neuron.axon.serve( self.model )
+        bittensor.axon.serve( self.model )
 
         # ---- Optimizer ----
         self.optimizer = torch.optim.SGD(self.model.parameters(), lr = self.config.miner.learning_rate, momentum=self.config.miner.momentum)
@@ -110,7 +110,7 @@ class Miner():
             for self.epoch in range(self.config.miner.n_epochs):
 
                 # ---- Poll until gradients ----
-                public_key, inputs_x, grads_dy, modality_x = bittensor.neuron.axon.gradients.get(block = True)
+                public_key, inputs_x, grads_dy, modality_x = bittensor.axon.gradients.get(block = True)
 
                 # ---- Backward Gradients ----
                 # TODO (const): batch normalization over the gradients for consistency.
@@ -127,18 +127,18 @@ class Miner():
                     self.model, self.optimizer = self.model_toolbox.load_model(self.config)
 
                 # ---- Step logs ----
-                logger.info('Step: {} \t Key: {} \t sum(W[:,0])', self.epoch, public_key, torch.sum(bittensor.neuron.metagraph.col()).item())
+                logger.info('Step: {} \t Key: {} \t sum(W[:,0])', self.epoch, public_key, torch.sum(bittensor.metagraph.col()).item())
             
                 # ---- Sync State ----
                 if (self.epoch + 1) % self.config.miner.sync_interval == 0:
 
                     # --- Display Epoch ----
-                    print(bittensor.neuron.axon.fullToString())
-                    print(bittensor.neuron.dendrite.fullToString())
-                    print(bittensor.neuron.metagraph)
+                    print(bittensor.axon.fullToString())
+                    print(bittensor.dendrite.fullToString())
+                    print(bittensor.metagraph)
                     
                     # ---- Sync metagrapn from chain ----
-                    bittensor.neuron.metagraph.sync() # Sync with the chain.
+                    bittensor.metagraph.sync() # Sync with the chain.
                     
                     # --- Save Model ----
                     self.model_toolbox.save_model(
